Move lexer main functionality to LexingBase superclass
